Improved NMAP speed, added direct pass
This allows you to pass a valid NMAP IP to NMAP directly.

It also changes the name of the incorrectly named `subnet` argument. A subnet is unrelated to the IP address directly.

Following this logic, there should be a subnet mask feature that gets the current subnet mask and returns it, then completing the nmap call using the following information:
https://en.wikipedia.org/wiki/Classless_Inter-Domain_Routing